Save Slot Update is Done
Letting player name the save slots, deleting save slots have been added. Also the player progress can be seen in the save slot via badges and background of the save slot. Background-save slot connection is done. Badge-save slot connection will be added in the future. Also interactibility of save slot is checked in Update function. Will be tested in the future it that situation is breaking the game or not.